+++ tests/Doctrine/Tests/ODM/PHPCR/Functional/EventManagerTest.php
@@ -2,13 +2,13 @@
 
 namespace Doctrine\Tests\ODM\PHPCR\Functional;
 
-use Doctrine\Common\EventArgs;
+use Doctrine\Common\Persistence\Event\LifecycleEventArgs;
+use Doctrine\Common\Persistence\Event\ManagerEventArgs;
+use Doctrine\ODM\PHPCR\Event\MoveEventArgs;
 
 use Doctrine\Tests\ODM\PHPCR\PHPCRFunctionalTestCase;
 use Doctrine\Tests\Models\CMS\CmsPage;
 use Doctrine\Tests\Models\CMS\CmsItem;
-use Doctrine\ODM\PHPCR\Event\PostFlushEventArgs;
-use Doctrine\ODM\PHPCR\Event\PreFlushEventArgs;
 
 class EventManagerTest extends PHPCRFunctionalTestCase
 {
@@ -130,9 +130,9 @@ class TestPersistenceListener
     public $pagePreMove = false;
     public $pagePostMove = false;
 
-    public function prePersist(EventArgs $e)
+    public function prePersist(LifecycleEventArgs $e)
     {
-        $document = $e->getDocument();
+        $document = $e->getObject();
         if ($document instanceof CmsPage){
             $this->pagePrePersist = true;
         } elseif ($document instanceof CmsItem){
@@ -140,9 +140,9 @@ class TestPersistenceListener
         }
     }
 
-    public function postPersist(EventArgs $e)
+    public function postPersist(LifecycleEventArgs $e)
     {
-        $document = $e->getDocument();
+        $document = $e->getObject();
         if ($document instanceof CmsPage){
             $this->pagePostPersist = true;
         } elseif ($document instanceof CmsItem){
@@ -150,13 +150,13 @@ class TestPersistenceListener
         }
     }
 
-    public function preUpdate(EventArgs $e)
+    public function preUpdate(LifecycleEventArgs $e)
     {
-        $document = $e->getDocument();
+        $document = $e->getObject();
         if (! $document instanceof CmsPage ){
             return;
         }
-        $dm = $e->getDocumentManager();
+        $dm = $e->getObjectManager();
 
         foreach ($document->getItems() as $item) {
             $dm->persist($item);
@@ -164,14 +164,14 @@ class TestPersistenceListener
         $this->preUpdate = true;
     }
 
-    public function postUpdate(EventArgs $e)
+    public function postUpdate(LifecycleEventArgs $e)
     {
         $this->postUpdate = true;
     }
 
-    public function preRemove(EventArgs $e)
+    public function preRemove(LifecycleEventArgs $e)
     {
-        $document = $e->getDocument();
+        $document = $e->getObject();
         if ($document instanceof CmsPage){
             $this->pagePreRemove = true;
         } elseif ($document instanceof CmsItem){
@@ -179,9 +179,9 @@ class TestPersistenceListener
         }
     }
 
-    public function postRemove(EventArgs $e)
+    public function postRemove(LifecycleEventArgs $e)
     {
-        $document = $e->getDocument();
+        $document = $e->getObject();
         if ($document instanceof CmsPage){
             $this->pagePostRemove = true;
         } elseif ($document instanceof CmsItem){
@@ -189,9 +189,9 @@ class TestPersistenceListener
         }
     }
 
-    public function preMove(EventArgs $e)
+    public function preMove(MoveEventArgs $e)
     {
-        $document = $e->getDocument();
+        $document = $e->getObject();
         if ($document instanceof CmsPage){
             $this->pagePreMove = true;
         } elseif ($document instanceof CmsItem){
@@ -199,9 +199,9 @@ class TestPersistenceListener
         }
     }
 
-    public function postMove(EventArgs $e)
+    public function postMove(LifecycleEventArgs $e)
     {
-        $document = $e->getDocument();
+        $document = $e->getObject();
         if ($document instanceof CmsPage){
             $this->pagePostMove = true;
         } elseif ($document instanceof CmsItem){
@@ -209,17 +209,17 @@ class TestPersistenceListener
         }
     }
 
-    public function onFlush(EventArgs $e)
+    public function onFlush(ManagerEventArgs $e)
     {
         $this->onFlush = true;
     }
 
-    public function postFlush(PostFlushEventArgs $e)
+    public function postFlush(ManagerEventArgs $e)
     {
         $this->postFlush = true;
     }
 
-    public function preFlush(PreFlushEventArgs $e)
+    public function preFlush(ManagerEventArgs $e)
     {
         $this->preFlush = true;
     }